requirements.txt
requests
beautifulsoup4
lxml
readability-lxml
elasticsearch
fastapi
uvicorn[standard]
loguru
